Create intermediate UserSession
